+++ firebase-functions/src/http-handlers/customer.js
